backend/requirements.txt
fastapi == 0.110.2
mysql-connector-python == 8.3.0
paho-mqtt == 2.0.0
urllib3 == 2.2.1
uvicorn == 0.29.0